+++ creme/linear_model/__init__.py
@@ -2,6 +2,7 @@
 from .fm import FMRegressor
 from .glm import LinearRegression
 from .glm import LogisticRegression
+from .glm import PoissonRegression
 from .pa import PAClassifier
 from .pa import PARegressor
 from .softmax import SoftmaxRegression
@@ -13,5 +14,6 @@ __all__ = [
     'LogisticRegression',
     'PAClassifier',
     'PARegressor',
+    'PoissonRegression',
     'SoftmaxRegression'
 ]