+++ CHANGELOG.md
@@ -18,7 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added the `AMSGrad` optimizer to the `optim` module
 - Added `RandomUnderSampler`, `RandomOverSampler`, and `RandomSampler` to the newly created `imblearn` module.
 - Added `PyTorch2CremeRegressor` and to the `compat` module
-- Added `Phishing` to `datasets`
+- Added `PoissonRegression` to `linear_model`
+- Added `Poisson` to `optim.losses`
 
 ### Fixed